edit `process_bin_WD1` input Change `sample_size` to allow for floats.

+++ packs/proc/processing_utils.py
@@ -424,7 +424,7 @@ def process_event_lazy_WD1(file_object  :  BinaryIO):
 
 def process_bin_WD1(file_path    :  str,
                     save_path    :  str,
-                    sample_size  :  int,
+                    sample_size  :  float,
                     overwrite    :  Optional[bool] = False,
                     print_mod    :  Optional[int] = -1):
 
@@ -446,7 +446,7 @@ def process_bin_WD1(file_path    :  str,
     ----------
         file_path    (str)   :  Path to binary file
         save_path    (str)   :  Path to saved file
-        sample_size  (int)   :  Size of each sample in an event (default 2 ns in the case of V1730B digitiser)
+        sample_size  (float)   :  Size of each sample in an event (default 2 ns in the case of V1730B digitiser)
         overwrite    (bool)  :  Boolean for overwriting pre-existing files
         print_mod    (int)   :  Readout frequency for number of events, -1 implies no readout
     Returns
